requirements.txt
addict==2.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.3
docker-pycreds==0.4.0
fair-esm==2.0.0
filelock==3.16.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.25.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
lit==18.1.8
lmdb==1.5.1
loguru==0.7.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==24.1
pandas==1.5.3
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.28.2
psutil==6.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rdkit==2024.3.5
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentry-sdk==2.14.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.5
transformers==4.44.2
triton==2.0.0
typing_extensions==4.12.2
unimol-tools==0.1.0.post4
urllib3==2.2.3
wandb==0.18.1
